Remove custom terminal implementation
