+++ extensions/gitpod-remote/tsconfig.json
@@ -3,7 +3,8 @@
 	"compilerOptions": {
 		"outDir": "./out",
 		"rootDir": "./src",
-		"esModuleInterop": true
+		"esModuleInterop": true,
+		"sourceMap": true,
 	},
 	"references": [
 		{ "path": "../gitpod-shared" }
