+++ extensions/gitpod-shared/package.json
@@ -7,6 +7,7 @@
   "engines": {
     "vscode": "^1.58.2"
   },
+  "main": "./out/extension.js",
   "scripts": {
     "prepare": "node scripts/inflate.js"
   },
